feat: Add HomePage hook
